requirements.txt
Django==5.1.1
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-filter==24.3